Refactor layer calculations into their model, add unit tests for extension methods
