+++ AI/Models/NeuralNetwork.Test/LayerExtensionsShould.cs
@@ -1,0 +1,32 @@
+﻿namespace NeuralNetwork.Test
+{
+    using Xunit;
+    using Xunit.Sdk;
+
+    public class LayerExtensionsShould
+    {
+        [Fact]
+        public void DeepCopyCorrectly()
+        {
+            // TODO
+        }
+
+        [Fact]
+        public void CloneWithNodeReferencesCorrectly()
+        {
+            // TODO
+        }
+
+        [Fact]
+        public void CloneNewWithWeightReferencesCorrectly()
+        {
+            // TODO
+        }
+
+        [Fact(Skip = "Debug only")]
+        public void SaveCorrectly()
+        {
+            // TODO
+        }
+    }
+}
